+++ Graph/src/OrientedGraph.cpp
@@ -15,13 +15,13 @@ OrientedGraph::~OrientedGraph()
 }
 void OrientedGraph::ReadFromFile(std::ifstream& file)
 {
-    unsigned int size, edgeSz;
+    unsigned long long size, edgeSz;
     file >> size;
     this->graph->InitializeNewGraph(size);
     file >> edgeSz;
-    for(int i = 0; i < edgeSz; ++i)
+    for(long long i = 0; i < edgeSz; ++i)
     {
-        unsigned int from, to;
+        unsigned long long from, to;
         file >> from;
         file >> to;
         this->AddEdge(from, to);
@@ -32,7 +32,7 @@ void OrientedGraph::WriteToFile(std::ofstream &file)
     std::vector<Edge*> edges = this->graph->GetAllEdges();
     file << this->Size() << std::endl;
     file << edges.size() << std::endl;
-    for(int i = 0; i < edges.size(); ++i)
+    for(long long i = 0; i < edges.size(); ++i)
         file << edges[i]->From << " " << edges[i]->To << std::endl;
 }
 void OrientedGraph::WriteToFile(std::string filename)
@@ -44,26 +44,26 @@ void OrientedGraph::WriteToFile(std::string filename)
 void OrientedGraph::RandomizeGraph(double probability)
 {
     this->InitializeNewGraph(this->Size());
-    for(int i = 0; i < this->Size(); ++i)
-        for(int j = 0; j < this->Size(); ++j)
+    for(long long i = 0; i < this->Size(); ++i)
+        for(long long j = 0; j < this->Size(); ++j)
         {
             if(j == i)
                 continue;
-            if(!(rand() % (int)(1.0 / probability)))
+            if(!(rand() % (long long)(1.0 / probability)))
             {
                 this->AddEdge(i, j);
             }
         }
 }
-void OrientedGraph::BFS(unsigned int source, std::function<void(unsigned int, unsigned int)> operation)
+void OrientedGraph::BFS(unsigned long long source, std::function<void(unsigned long long, unsigned long long)> operation)
 {
     enum Color{White, Grey, Black};
     std::vector<Color> vertexes (this->Size(), Color::White);
-    std::queue<unsigned int> q;
+    std::queue<unsigned long long> q;
     q.push(source);
     while(!q.empty())
     {
-        unsigned int u = q.front();
+        unsigned long long u = q.front();
         q.pop();
         if(vertexes[u] == Color::Black)
             continue;
