+++ Graph/src/UnorientedGraph.cpp
@@ -12,13 +12,13 @@ UnorientedGraph::~UnorientedGraph()
 
 void UnorientedGraph::ReadFromFile(std::ifstream &file)
 {
-    unsigned int size, edgeSz;
+    unsigned long long size, edgeSz;
     file >> size;
     this->InitializeNewGraph(size);
     file >> edgeSz;
-    for(int i = 0; i < edgeSz; ++i)
+    for(long long i = 0; i < edgeSz; ++i)
     {
-        unsigned int from, to;
+        unsigned long long from, to;
         file >> from;
         file >> to;
         this->AddEdge(from, to);
@@ -35,18 +35,18 @@ void UnorientedGraph::WriteToFile(std::ofstream &file)
     std::vector<Edge*> edges = this->graph->GetAllEdges();
     file << this->Size() << std::endl;
     file << edges.size() << std::endl;
-    for(int i = 0; i < edges.size(); ++i)
+    for(long long i = 0; i < edges.size(); ++i)
         file << edges[i]->From << " " << edges[i]->To << std::endl;
 }
 void UnorientedGraph::RandomizeGraph(double probability)
 {
     this->InitializeNewGraph(this->Size());
-    for(int i = 0; i < this->Size() - 1; ++i)
-        for(int j = i; j < this->Size(); ++j)
+    for(long long i = 0; i < this->Size() - 1; ++i)
+        for(long long j = i; j < this->Size(); ++j)
         {
             if(j == i)
                 continue;
-            if(!(rand() % (int)(1.0 / probability)))
+            if(!(rand() % (long long)(1.0 / probability)))
             {
                 this->AddEdge(i, j);
             }
@@ -55,10 +55,10 @@ void UnorientedGraph::RandomizeGraph(double probability)
 void UnorientedGraph::NormalizeEdges()
 {
     std::vector<Edge*> edges = this->graph->GetAllEdges();
-    for(int i = 0; i < edges.size(); ++i)
+    for(long long i = 0; i < edges.size(); ++i)
         if(edges[i]->From > edges[i]->To)
         {
-            unsigned int temp = edges[i]->From;
+            unsigned long long temp = edges[i]->From;
             edges[i]->From = edges[i]->To;
             edges[i]->To = temp;
         }
